improvement(feature list): control features from one array, don't hard code them in a reducer
The list of features were in an array, but they were also hardcoded in a reducer. This meant that to
add a feature, you would have had to modify that reducer, as well as the list of arrays. This fixes
that problem. It also makes the code uglier - I couldn't come up with a generic internal reducer for
both creating the initial object and running the outer reducer. Hope to refactor. See you in 2030.